mcp-documents/requirements.txt
mcp[cli]>=1.0.0
pymupdf>=1.24.0
Pillow>=10.0.0
minio>=7.2.0
structlog>=24.0.0
httpx>=0.27.0
